+++ benchmarks/bm_runner.py
@@ -61,7 +61,7 @@ def _check_requirements(package: str) -> None:
 
 def _prep_data_gen_env() -> None:
     """Create or access a separate, unchanging environment for generating test data."""
-    python_version = "3.10"
+    python_version = "3.12"
     data_gen_var = "DATA_GEN_PYTHON"
     if data_gen_var in environ:
         _echo("Using existing data generation environment.")
@@ -279,7 +279,7 @@ class Custom(_SubParserGenerator):
     description = (
         "Run ASV with the input **ASV sub-command**, without any preset "
         "arguments - must all be supplied by the user. So just like running "
-        "ASV manually, with the convenience of re-using the runner's "
+        "ASV manually, with the convenience of reusing the runner's "
         "scripted setup steps."
     )
     epilog = "e.g. python bm_runner.py custom continuous a1b23d4 HEAD --quick"