+++ esmf_regrid/schemes.py
@@ -19,7 +19,7 @@ except ImportError as exc:
     except ImportError:
         raise exc
 
-from esmf_regrid import Constants, check_method
+from esmf_regrid import Constants, check_method, esmpy
 from esmf_regrid.esmf_regridder import GridInfo, RefinedGridInfo, Regridder
 from esmf_regrid.experimental.unstructured_regrid import MeshInfo
 
@@ -32,6 +32,12 @@ __all__ = [
     "ESMFNearestRegridder",
     "regrid_rectilinear_to_rectilinear",
 ]
+
+STANDARD_BILINEAR_EXTRAP_ARGS = {
+    "extrap_method": esmpy.ExtrapMethod.NEAREST_IDAVG,
+    "extrap_num_src_pnts": 2,
+    "extrap_dist_exponent": 1,
+}
 
 
 def _get_coord(cube, axis):
@@ -289,6 +295,42 @@ def _regrid_along_dims(data, regridder, dims, num_out_dims, mdtol):
         dims = [dims[0] + 1, dims[0]]
     result = np.moveaxis(result, standard_out_dims, dims)
     return result
+
+
+def _check_esmf_args(kwargs):
+    if kwargs is not None:
+        if not isinstance(kwargs, dict):
+            raise TypeError(f"Expected `esmf_args` to be a dict, got a {type(kwargs)}.")
+        invalid_kwargs = [
+            "filename",
+            "norm_type",
+            "rh_filename",
+            "regrid_method",
+            "src_mask_values",
+            "dst_mask_values",
+            "factors",
+            "src_frac_field",
+            "dst_frac_field",
+        ]
+        valid_kwargs = [
+            "pole_method",
+            "regrid_pole_npoints",
+            "line_type",
+            "extrap_method",
+            "extrap_num_src_pnts",
+            "extrap_dist_exponent",
+            "extrap_num_levels",
+            "unmapped_action",
+            "ignore_degenerate",
+            "large_file",
+        ]
+        for kwarg in kwargs:
+            if kwarg in invalid_kwargs:
+                msg = f"`esmpy.Regrid` argument `{kwarg}` cannot be controlled by `esmf_args`."
+                raise ValueError(msg)
+            if kwarg not in valid_kwargs:
+                msg = f"`{kwarg}` is not a valid argument for `esmpy.Regrid`."
+                raise ValueError(msg)
 
 
 def _map_complete_blocks(
@@ -549,11 +591,12 @@ def _regrid_rectilinear_to_rectilinear__prepare(
     tgt_resolution=None,
     src_mask=None,
     tgt_mask=None,
+    esmf_args=None,
 ):
     """First (setup) part of 'regrid_rectilinear_to_rectilinear'.
 
     Check inputs and calculate the sparse regrid matrix and related info.
-    The 'regrid info' returned can be re-used over many 2d slices.
+    The 'regrid info' returned can be reused over many 2d slices.
 
     """
     tgt_x = _get_coord(tgt_grid_cube, "x")
@@ -570,8 +613,14 @@ def _regrid_rectilinear_to_rectilinear__prepare(
     srcinfo = _make_gridinfo(src_grid_cube, method, src_resolution, src_mask)
     tgtinfo = _make_gridinfo(tgt_grid_cube, method, tgt_resolution, tgt_mask)
 
+    _check_esmf_args(esmf_args)
+
     regridder = Regridder(
-        srcinfo, tgtinfo, method=method, precomputed_weights=precomputed_weights
+        srcinfo,
+        tgtinfo,
+        method=method,
+        esmf_args=esmf_args,
+        precomputed_weights=precomputed_weights,
     )
 
     regrid_info = RegridInfo(
@@ -633,11 +682,12 @@ def _regrid_unstructured_to_rectilinear__prepare(
     tgt_resolution=None,
     src_mask=None,
     tgt_mask=None,
+    esmf_args=None,
 ):
     """First (setup) part of 'regrid_unstructured_to_rectilinear'.
 
     Check inputs and calculate the sparse regrid matrix and related info.
-    The 'regrid info' returned can be re-used over many 2d slices.
+    The 'regrid info' returned can be reused over many 2d slices.
 
     """
     grid_x = _get_coord(target_grid_cube, "x")
@@ -650,8 +700,14 @@ def _regrid_unstructured_to_rectilinear__prepare(
     meshinfo = _make_meshinfo(src_mesh_cube, method, src_mask, "source")
     gridinfo = _make_gridinfo(target_grid_cube, method, tgt_resolution, tgt_mask)
 
+    _check_esmf_args(esmf_args)
+
     regridder = Regridder(
-        meshinfo, gridinfo, method=method, precomputed_weights=precomputed_weights
+        meshinfo,
+        gridinfo,
+        method=method,
+        esmf_args=esmf_args,
+        precomputed_weights=precomputed_weights,
     )
 
     regrid_info = RegridInfo(
@@ -717,11 +773,12 @@ def _regrid_rectilinear_to_unstructured__prepare(
     src_mask=None,
     tgt_mask=None,
     tgt_location=None,
+    esmf_args=None,
 ):
     """First (setup) part of 'regrid_rectilinear_to_unstructured'.
 
     Check inputs and calculate the sparse regrid matrix and related info.
-    The 'regrid info' returned can be re-used over many 2d slices.
+    The 'regrid info' returned can be reused over many 2d slices.
 
     """
     grid_x = _get_coord(src_grid_cube, "x")
@@ -744,8 +801,14 @@ def _regrid_rectilinear_to_unstructured__prepare(
     )
     gridinfo = _make_gridinfo(src_grid_cube, method, src_resolution, src_mask)
 
+    _check_esmf_args(esmf_args)
+
     regridder = Regridder(
-        gridinfo, meshinfo, method=method, precomputed_weights=precomputed_weights
+        gridinfo,
+        meshinfo,
+        method=method,
+        esmf_args=esmf_args,
+        precomputed_weights=precomputed_weights,
     )
 
     regrid_info = RegridInfo(
@@ -814,11 +877,12 @@ def _regrid_unstructured_to_unstructured__prepare(
     tgt_mask=None,
     src_location=None,
     tgt_location=None,
+    esmf_args=None,
 ):
     """First (setup) part of 'regrid_unstructured_to_unstructured'.
 
     Check inputs and calculate the sparse regrid matrix and related info.
-    The 'regrid info' returned can be re-used over many 2d slices.
+    The 'regrid info' returned can be reused over many 2d slices.
 
     """
     if isinstance(tgt_cube_or_mesh, MeshXY):
@@ -837,11 +901,14 @@ def _regrid_unstructured_to_unstructured__prepare(
         tgt_cube_or_mesh, method, tgt_mask, "target", location=tgt_location
     )
 
+    _check_esmf_args(esmf_args)
+
     regridder = Regridder(
         src_meshinfo,
         tgt_meshinfo,
         method=method,
         precomputed_weights=precomputed_weights,
+        esmf_args=esmf_args,
     )
 
     regrid_info = RegridInfo(
@@ -962,7 +1029,12 @@ class ESMFAreaWeighted:
     """
 
     def __init__(
-        self, mdtol=0, use_src_mask=False, use_tgt_mask=False, tgt_location="face"
+        self,
+        mdtol=0,
+        use_src_mask=False,
+        use_tgt_mask=False,
+        tgt_location="face",
+        esmf_args=None,
     ):
         """Area-weighted scheme for regridding between rectilinear grids.
 
@@ -985,6 +1057,10 @@ class ESMFAreaWeighted:
         tgt_location : str or None, default="face"
             Either "face" or "node". Describes the location for data on the mesh
             if the target is not a :class:`~iris.cube.Cube`.
+        esmf_args : dict, optional
+            A dictionary of arguments to pass to ESMF. For the sake of reference, These
+            arguments are recorded as a property of this regridder and are stored when
+            the regridder is saved .
 
         """
         if not (0 <= mdtol <= 1):
@@ -997,10 +1073,21 @@ class ESMFAreaWeighted:
         self.use_src_mask = use_src_mask
         self.use_tgt_mask = use_tgt_mask
         self.tgt_location = "face"
+        if esmf_args is None:
+            esmf_args = {}
+        _check_esmf_args(esmf_args)
+        self.esmf_args = esmf_args
 
     def __repr__(self):
         """Return a representation of the class."""
-        return f"ESMFAreaWeighted(mdtol={self.mdtol})"
+        result = (
+            f"ESMFAreaWeighted("
+            f"mdtol={self.mdtol}, "
+            f"use_src_mask={self.use_src_mask}, "
+            f"use_tgt_mask={self.use_tgt_mask}, "
+            f"esmf_args={self.esmf_args})"
+        )
+        return result
 
     def regridder(
         self,
@@ -1011,6 +1098,7 @@ class ESMFAreaWeighted:
         use_src_mask=None,
         use_tgt_mask=None,
         tgt_location="face",
+        esmf_args=None,
     ):
         """Create regridder to perform regridding from ``src_grid`` to ``tgt_grid``.
 
@@ -1039,6 +1127,10 @@ class ESMFAreaWeighted:
         tgt_location : str or None, default="face"
             Either "face" or "node". Describes the location for data on the mesh
             if the target is not a :class:`~iris.cube.Cube`.
+        esmf_args : dict, optional
+            A dictionary of arguments to pass to ESMF. For the sake of reference, These
+            arguments are recorded as a property of this regridder and are stored when
+            the regridder is saved .
 
 
         Returns
@@ -1059,6 +1151,9 @@ class ESMFAreaWeighted:
             use_src_mask = self.use_src_mask
         if use_tgt_mask is None:
             use_tgt_mask = self.use_tgt_mask
+        if esmf_args is None:
+            esmf_args = self.esmf_args
+
         if tgt_location is not None and tgt_location != "face":
             e_msg = "For area weighted regridding, target location must be 'face'."
             raise ValueError(e_msg)
@@ -1071,6 +1166,7 @@ class ESMFAreaWeighted:
             use_src_mask=use_src_mask,
             use_tgt_mask=use_tgt_mask,
             tgt_location="face",
+            esmf_args=esmf_args,
         )
 
 
@@ -1083,7 +1179,13 @@ class ESMFBilinear:
     """
 
     def __init__(
-        self, mdtol=0, use_src_mask=False, use_tgt_mask=False, tgt_location=None
+        self,
+        mdtol=0,
+        use_src_mask=False,
+        use_tgt_mask=False,
+        tgt_location=None,
+        extrapolate_gaps=False,
+        esmf_args=None,
     ):
         """Area-weighted scheme for regridding between rectilinear grids.
 
@@ -1102,6 +1204,16 @@ class ESMFBilinear:
         tgt_location : str or None, default=None
             Either "face" or "node". Describes the location for data on the mesh
             if the target is not a :class:`~iris.cube.Cube`.
+        extrapolate_gaps : bool, default=False
+            Use a standard set of ESMF arguments for extrapolation which achieves
+            continuity with bilinear regridding. Useful for situations where gaps
+            between cells would be masked. Note: this overwrites any arguments passed
+            to ``esmf_args`` for the keywords "extrap_method", "extrap_num_src_pnts"
+            or "extrap_dist_exponent".
+        esmf_args : dict, optional
+            A dictionary of arguments to pass to ESMF. For the sake of reference, These
+            arguments are recorded as a property of this regridder and are stored when
+            the regridder is saved .
 
         """
         if not (0 <= mdtol <= 1):
@@ -1111,10 +1223,23 @@ class ESMFBilinear:
         self.use_src_mask = use_src_mask
         self.use_tgt_mask = use_tgt_mask
         self.tgt_location = tgt_location
+        if esmf_args is None:
+            esmf_args = {}
+        if extrapolate_gaps:
+            esmf_args.update(STANDARD_BILINEAR_EXTRAP_ARGS)
+        _check_esmf_args(esmf_args)
+        self.esmf_args = esmf_args
 
     def __repr__(self):
         """Return a representation of the class."""
-        return f"ESMFBilinear(mdtol={self.mdtol})"
+        result = (
+            f"ESMFBilinear("
+            f"mdtol={self.mdtol}, "
+            f"use_src_mask={self.use_src_mask}, "
+            f"use_tgt_mask={self.use_tgt_mask}, "
+            f"esmf_args={self.esmf_args})"
+        )
+        return result
 
     def regridder(
         self,
@@ -1123,6 +1248,8 @@ class ESMFBilinear:
         use_src_mask=None,
         use_tgt_mask=None,
         tgt_location=None,
+        extrapolate_gaps=False,
+        esmf_args=None,
     ):
         """Create regridder to perform regridding from ``src_grid`` to ``tgt_grid``.
 
@@ -1142,6 +1269,16 @@ class ESMFBilinear:
         tgt_location : str or None, default=None
             Either "face" or "node". Describes the location for data on the mesh
             if the target is not a :class:`~iris.cube.Cube`.
+        extrapolate_gaps : bool, default=False
+            Use a standard set of ESMF arguments for extrapolation which achieves
+            continuity with bilinear regridding. Useful for situations where gaps
+            between cells would be masked. Note: this overwrites any arguments passed
+            to ``esmf_args`` for the keywords "extrap_method", "extrap_num_src_pnts"
+            or "extrap_dist_exponent".
+        esmf_args : dict, optional
+            A dictionary of arguments to pass to ESMF. For the sake of reference, These
+            arguments are recorded as a property of this regridder and are stored when
+            the regridder is saved .
 
         Returns
         -------
@@ -1163,6 +1300,10 @@ class ESMFBilinear:
             use_tgt_mask = self.use_tgt_mask
         if tgt_location is None:
             tgt_location = self.tgt_location
+        if esmf_args is None:
+            esmf_args = self.esmf_args
+        if extrapolate_gaps:
+            esmf_args = STANDARD_BILINEAR_EXTRAP_ARGS
         return ESMFBilinearRegridder(
             src_grid,
             tgt_grid,
@@ -1170,6 +1311,7 @@ class ESMFBilinear:
             use_src_mask=use_src_mask,
             use_tgt_mask=use_tgt_mask,
             tgt_location=tgt_location,
+            esmf_args=esmf_args,
         )
 
 
@@ -1198,7 +1340,13 @@ class ESMFNearest:
     the same equivalent space will behave the same.
     """
 
-    def __init__(self, use_src_mask=False, use_tgt_mask=False, tgt_location=None):
+    def __init__(
+        self,
+        use_src_mask=False,
+        use_tgt_mask=False,
+        tgt_location=None,
+        esmf_args=None,
+    ):
         """Nearest neighbour scheme for regridding between rectilinear grids.
 
         Parameters
@@ -1212,14 +1360,28 @@ class ESMFNearest:
         tgt_location : str or None, default=None
             Either "face" or "node". Describes the location for data on the mesh
             if the target is not a :class:`~iris.cube.Cube`.
+        esmf_args : dict, optional
+            A dictionary of arguments to pass to ESMF. For the sake of reference, These
+            arguments are recorded as a property of this regridder and are stored when
+            the regridder is saved .
         """
         self.use_src_mask = use_src_mask
         self.use_tgt_mask = use_tgt_mask
         self.tgt_location = tgt_location
+        if esmf_args is None:
+            esmf_args = {}
+        _check_esmf_args(esmf_args)
+        self.esmf_args = esmf_args
 
     def __repr__(self):
         """Return a representation of the class."""
-        return "ESMFNearest()"
+        result = (
+            f"ESMFNearest("
+            f"use_src_mask={self.use_src_mask}, "
+            f"use_tgt_mask={self.use_tgt_mask}, "
+            f"esmf_args={self.esmf_args})"
+        )
+        return result
 
     def regridder(
         self,
@@ -1228,6 +1390,7 @@ class ESMFNearest:
         use_src_mask=None,
         use_tgt_mask=None,
         tgt_location=None,
+        esmf_args=None,
     ):
         """Create regridder to perform regridding from ``src_grid`` to ``tgt_grid``.
 
@@ -1247,6 +1410,10 @@ class ESMFNearest:
         tgt_location : str or None, default=None
             Either "face" or "node". Describes the location for data on the mesh
             if the target is not a :class:`~iris.cube.Cube`.
+        esmf_args : dict, optional
+            A dictionary of arguments to pass to ESMF. For the sake of reference, These
+            arguments are recorded as a property of this regridder and are stored when
+            the regridder is saved .
 
         Returns
         -------
@@ -1268,12 +1435,15 @@ class ESMFNearest:
             use_tgt_mask = self.use_tgt_mask
         if tgt_location is None:
             tgt_location = self.tgt_location
+        if esmf_args is None:
+            esmf_args = self.esmf_args
         return ESMFNearestRegridder(
             src_grid,
             tgt_grid,
             use_src_mask=use_src_mask,
             use_tgt_mask=use_tgt_mask,
             tgt_location=tgt_location,
+            esmf_args=esmf_args,
         )
 
 
@@ -1289,6 +1459,7 @@ class _ESMFRegridder:
         use_src_mask=False,
         use_tgt_mask=False,
         tgt_location=None,
+        esmf_args=None,
         **kwargs,
     ):
         """Create regridder for conversions between ``src`` and ``tgt``.
@@ -1335,6 +1506,11 @@ class _ESMFRegridder:
             raise ValueError(msg.format(mdtol))
         self.mdtol = mdtol
         self.method = method
+
+        if esmf_args is None:
+            esmf_args = {}
+        self.esmf_args = esmf_args
+        kwargs["esmf_args"] = self.esmf_args
 
         self.src_mask = _get_mask(src, use_src_mask)
         kwargs["src_mask"] = self.src_mask
@@ -1466,6 +1642,7 @@ class ESMFAreaWeightedRegridder(_ESMFRegridder):
         use_src_mask=False,
         use_tgt_mask=False,
         tgt_location="face",
+        esmf_args=None,
     ):
         """Create regridder for conversions between ``src`` and ``tgt``.
 
@@ -1503,6 +1680,10 @@ class ESMFAreaWeightedRegridder(_ESMFRegridder):
         tgt_location : str or None, default="face"
             Either "face" or "node". Describes the location for data on the mesh
             if the target is not a :class:`~iris.cube.Cube`.
+        esmf_args : dict, optional
+            A dictionary of arguments to pass to ESMF. For the sake of reference, These
+            arguments are recorded as a property of this regridder and are stored when
+            the regridder is saved .
 
         Raises
         ------
@@ -1530,6 +1711,7 @@ class ESMFAreaWeightedRegridder(_ESMFRegridder):
             mdtol=mdtol,
             precomputed_weights=precomputed_weights,
             tgt_location="face",
+            esmf_args=esmf_args,
             **kwargs,
         )
 
@@ -1546,6 +1728,8 @@ class ESMFBilinearRegridder(_ESMFRegridder):
         use_src_mask=False,
         use_tgt_mask=False,
         tgt_location=None,
+        extrapolate_gaps=False,
+        esmf_args=None,
     ):
         """Create regridder for conversions between ``src`` and ``tgt``.
 
@@ -1574,6 +1758,16 @@ class ESMFBilinearRegridder(_ESMFRegridder):
         tgt_location : str or None, default=None
             Either "face" or "node". Describes the location for data on the mesh
             if the target is not a :class:`~iris.cube.Cube`.
+        extrapolate_gaps : bool, default=False
+            Use a standard set of ESMF arguments for extrapolation which achieves
+            continuity with bilinear regridding. Useful for situations where gaps
+            between cells would be masked. Note: this overwrites any arguments passed
+            to ``esmf_args`` for the keywords "extrap_method", "extrap_num_src_pnts"
+            or "extrap_dist_exponent".
+        esmf_args : dict, optional
+            A dictionary of arguments to pass to ESMF. For the sake of reference, These
+            arguments are recorded as a property of this regridder and are stored when
+            the regridder is saved .
 
         Raises
         ------
@@ -1581,6 +1775,10 @@ class ESMFBilinearRegridder(_ESMFRegridder):
             If ``use_src_mask`` or ``use_tgt_mask`` are True while the masks on ``src``
             or ``tgt`` respectively are not constant over non-horizontal dimensions.
         """
+        if esmf_args is None:
+            esmf_args = {}
+        if extrapolate_gaps:
+            esmf_args.update(STANDARD_BILINEAR_EXTRAP_ARGS)
         super().__init__(
             src,
             tgt,
@@ -1590,6 +1788,7 @@ class ESMFBilinearRegridder(_ESMFRegridder):
             use_src_mask=use_src_mask,
             use_tgt_mask=use_tgt_mask,
             tgt_location=tgt_location,
+            esmf_args=esmf_args,
         )
 
 
@@ -1604,6 +1803,7 @@ class ESMFNearestRegridder(_ESMFRegridder):
         use_src_mask=False,
         use_tgt_mask=False,
         tgt_location=None,
+        esmf_args=None,
     ):
         """Create regridder for conversions between ``src`` and ``tgt``.
 
@@ -1626,6 +1826,10 @@ class ESMFNearestRegridder(_ESMFRegridder):
         tgt_location : str or None, default=None
             Either "face" or "node". Describes the location for data on the mesh
             if the target is not a :class:`~iris.cube.Cube`.
+        esmf_args : dict, optional
+            A dictionary of arguments to pass to ESMF. For the sake of reference, These
+            arguments are recorded as a property of this regridder and are stored when
+            the regridder is saved .
 
         Raises
         ------
@@ -1642,4 +1846,5 @@ class ESMFNearestRegridder(_ESMFRegridder):
             use_src_mask=use_src_mask,
             use_tgt_mask=use_tgt_mask,
             tgt_location=tgt_location,
+            esmf_args=esmf_args,
         )