+++ esmf_regrid/tests/unit/experimental/io/test_round_tripping.py
@@ -10,7 +10,9 @@ from esmf_regrid import (
     ESMFAreaWeightedRegridder,
     ESMFBilinear,
     ESMFNearest,
+    esmpy,
 )
+from esmf_regrid.esmf_regridder import ESMF_NO_VERSION
 from esmf_regrid.experimental.io import load_regridder, save_regridder
 from esmf_regrid.experimental.unstructured_scheme import (
     GridToMeshESMFRegridder,
@@ -456,7 +458,14 @@ def test_generic_regridder(tmp_path, src_type, tgt_type, scheme):
     elif tgt_type == "mesh":
         tgt = _gridlike_mesh_cube(n_lons_tgt, n_lats_tgt)
 
-    original_rg = scheme().regridder(src, tgt)
+    esmf_args = {
+        "unmapped_action": esmpy.UnmappedAction.ERROR,
+        "ignore_degenerate": False,
+        "line_type": esmpy.LineType.CART,
+        "large_file": True,
+    }
+
+    original_rg = scheme().regridder(src, tgt, esmf_args=esmf_args)
     filename = tmp_path / "regridder.nc"
     save_regridder(original_rg, filename)
     loaded_rg = load_regridder(str(filename))
@@ -469,6 +478,7 @@ def test_generic_regridder(tmp_path, src_type, tgt_type, scheme):
         assert original_rg.src_resolution == loaded_rg.src_resolution
         assert original_rg.tgt_resolution == loaded_rg.tgt_resolution
     assert original_rg.mdtol == loaded_rg.mdtol
+    assert original_rg.esmf_args == loaded_rg.esmf_args
 
 
 @pytest.mark.parametrize(
@@ -548,3 +558,18 @@ def test_generic_regridder_resolution(tmp_path, scheme):
     assert loaded_rg.regridder.src.resolution == src_resolution
     assert loaded_rg.tgt_resolution == tgt_resolution
     assert loaded_rg.regridder.tgt.resolution == tgt_resolution
+
+
+def test_precomputed(tmp_path):
+    """Test save/load round tripping for regridders made with precomputed weights."""
+    original_rg, src = _make_grid_to_mesh_regridder(
+        method=None, regridder=ESMFAreaWeightedRegridder
+    )
+    tgt = original_rg._tgt
+    weights = original_rg.regridder.weight_matrix
+    original_rg = ESMFAreaWeightedRegridder(src, tgt, precomputed_weights=weights)
+    assert original_rg.regridder.esmf_version == ESMF_NO_VERSION
+    filename = tmp_path / "regridder.nc"
+    save_regridder(original_rg, filename)
+    loaded_rg = load_regridder(str(filename))
+    assert original_rg.regridder.esmf_version == loaded_rg.regridder.esmf_version