+++ esmf_regrid/tests/unit/experimental/unstructured_scheme/test_MeshToGridESMFRegridder.py
@@ -263,7 +263,7 @@ def test_default_mdtol():
     assert rg_bi.mdtol == 0
 
 
-@pytest.mark.xfail()
+@pytest.mark.xfail
 def test_mistmatched_mesh():
     """Test the calling of :class:`esmf_regrid.experimental.unstructured_scheme.MeshToGridESMFRegridder`.
 
@@ -289,8 +289,7 @@ def test_mistmatched_mesh():
     with pytest.raises(ValueError) as excinfo:
         _ = rg(other_loc)
     expected_message = (
-        "The given cube is not defined on a the same "
-        "mesh location as this regridder."
+        "The given cube is not defined on the same mesh location as this regridder."
     )
     assert expected_message in str(excinfo.value)
     with pytest.raises(ValueError) as excinfo: