+++ noxfile.py
@@ -21,11 +21,14 @@ nox.options.reuse_existing_virtualenvs = False
 #: Name of the package to test.
 PACKAGE = "esmf_regrid"
 
-#: GHA-CI environment variable hook.
-PY_VER = os.environ.get("PY_VER", ["3.10", "3.11", "3.12"])
+#: Python versions we can run sessions under
+_PY_VERSIONS_ALL = ["3.11", "3.12", "3.13"]
 
 #: GHA-CI environment variable hook.
-COVERAGE = os.environ.get("COVERAGE", False)
+PY_VER = os.environ.get("PY_VER", _PY_VERSIONS_ALL)
+
+#: GHA-CI environment variable hook.
+COVERAGE = os.environ.get("COVERAGE", "")
 
 #: GHA-CI environment variable hook.
 #: If you change the IRIS_SOURCE here you will also need to change it in