+++ CHANGELOG.md
@@ -6,6 +6,40 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.12.1] - 2025-04-09
+
+### Fixed
+
+- [PR#486](https://github.com/SciTools/iris-esmf-regrid/pull/486)
+  Fixed bug causing the esmf argument `pole_method` to be rejected.
+  [@stephenworsley](https://github.com/stephenworsley)
+
+- [PR#479](https://github.com/SciTools/iris-esmf-regrid/pull/479)
+  Fixed bug which caused loading to be backwards incompatible.
+  [@stephenworsley](https://github.com/stephenworsley)
+
+## [0.12] - 2025-03-17
+
+### Fixed
+
+- [PR#462](https://github.com/SciTools/iris-esmf-regrid/pull/462)
+  Fixed bug when saving regridders made with precomputed weights.
+  [@stephenworsley](https://github.com/stephenworsley)
+
+### Added
+
+- [PR#464](https://github.com/SciTools/iris-esmf-regrid/pull/464)
+  Add support for Python 3.13.
+  [@stephenworsley](https://github.com/stephenworsley)
+- [PR#396](https://github.com/SciTools/iris-esmf-regrid/pull/396)
+  Added controls for passing arguments to ESMF.
+  [@stephenworsley](https://github.com/stephenworsley)
+
+### Removed
+
+- [PR#464](https://github.com/SciTools/iris-esmf-regrid/pull/464)
+  Remove support for Python 3.10.
+  [@stephenworsley](https://github.com/stephenworsley)
 
 ## [0.11] - 2024-07-31